Fix to prevent deleting required email
The default behavior of Django Allauth is allowing for a user's last and
primary email to be deleted, which will essentially lock the account. It looks
like there's a mis-match in settings, where it will only prevent deleting the
email if settings require using email address to login. The added tests verify
the current behavior with the stock account adapter, and the behavior we want
with our custom account adapter; which is, a user cannot delete their primary
email address until they designate another address as their primary.