Remove unused code in advance of new ICE interface
Lots of code relating to the Strain model and interfacing with ICE is removed
here, to help prepare the way for replacing the ICE interactions entirely with
use of API keys per-user, instead of a global admin-level HMAC key.

The interface in the Django admin to search strains as if another EDD user is
removed, as it wasn't working anyway, and the code it uses is about to be
replaced. Also the management command to assign legacy links inside EDD for
strains that have no ICE URL, using code about to be replaced, is removed.
Since the command was already run on all the ESE deployments of EDD, it no
longer needs to exist, as it won't do anything.