Add AppLinks to user profiles
As a pre-requisite to moving to multiple-ICE connections in a single EDD
deployment, we need a way to store credentials for each user to an arbitrary
ICE instance. The AppLinks feature does that, allowing each user to submit an
API token, client ID, and URL for zero or more ICE servers. These credentials
can then be used by upcoming changes to the ICE searching code.