Update ICE search interface for multi-ICE
With the addition of AppLinks, EDD can now provide a search interface to ICE
that is per-user, and allows individual users to link to an arbitrary ICE
instance. Any user without a valid ICE AppLink should see a disabled search
box for finding Strains, with instructions to set up an ICE connection.

Part of this update is a simplification of the Strain model. Removing the
unused parts of the Strain model both makes it easier to maintain and later
extend, and also helps to find places in code where the Strain model is used,
and ensure that every place is updated to the new search model.

All autocompletes using the older JQuery-based widgets are removed, except
for the ones on the SBML export page.