Replace all hard-coded WebSocket URLs
This is a long-time TODO item. Django makes use of a system where URLs to views
are refered to by names, instead of hard coding a path. The support for this on
WebSocket URLs is poor, and as a result EDD has used URLs hard coded into the
application. With a bit of custom coding, the Django support for getting URLs
to normal views can be adapted to WebSocket async views.