Fix static files handling
Adding a collectstatic step to run_tests.sh, since the test environment
overrides the `STORAGES` setting, the test static files storage may differ from
the development static files. As it is not guaranteed that the static file
collection will have run for the test environment, enforce running it before
starting tests.

The Dockerfile build also does a collectstatic step, without hashing and
manifests, to aggregate all the static files into one place in the image. This
action was using the older `STATICFILES_STORAGE` setting, instead of the new
`STORAGES` setting. This is updated in the build settings now to use the
correct name.