Update http service to use Gunicorn+uvicorn
In some testing I've done in local development in changing the proxy to EDD
services has revealed that ASGI server(s) like Daphne and Gunicorn run with a
uvicorn worker, are capable of handling both HTTP and WebSocket traffic. This
update switches the "main" `http` service to run Gunicorn with a uvicorn
worker. Originally, I tried with only Daphne, but it does not have as much
production-hardening, and doesn't support code reload with a hangup signal.

Once this is deployed to production, configs can be changed to only have two
containers using the EDD image, one for handling both HTTP and WebSocket
traffic, and the other to run Celery for background workers. The existing
`websocket` service can be removed, and the complicated proxy setup splitting
traffic between the two services can revert to a simple proxy.

After some time, we can also remove the Daphne dependency, and likely a few
others that were used by Daphne (gevent?).