Remove SBML export code
The code for SBML exports is among the oldest parts of EDD, and the most broken
as EDD's data model evolved while it remained stagnant. The button to export
study data to SBML doesn't currently work in production, and has not worked for
some time. The code had minimal testing, and is rife with perl-isms. As one of
the last remaining blockers to eliminating the non-Bootstrap5 front-end code,
we're better off removing this and implementing it properly when/if it becomes
important again.

Some lessons-learned on how *not* to do things from this implementation:
- Identifiers used in an SBML template file are unique to that template, and
  are not global IDs to tie to e.g. MeasurementType objects
- The SBML format has updated from the format used at the time; the hacky
  ways of embedding data in the file should use the format's explicit way to
  store that data