Fix Description page dropzone
In the switch to the "Experiment Setup" feature, the dropzone on the Study
Description page to add more lines, did not update completely. It was sending
data to the new view, but was expecting responses like the old "Experiment
Description" views.

This is fixed, and the unused front-end code is deleted.