Add custom Uvicorn worker class
The default Uvicorn worker looks for a "lifespan" protocol in the ASGI app,
which is not currently provided in Django Channels (and seems to be optional?).
This adds spurious warnings at server startup. The worker can be started
without trying the "lifespan" protocol, and the custom worker launches with
settings to turn it off.

See: https://github.com/encode/uvicorn/issues/709