Fix another migration failure
Switching to the updated Strain model tries to copy the registry ID and URL to
the new external ID and URL fields, but the external URL is annotated as
unique. While the old model is not annotated as unique, but the code was
generally treating it that way. Updating the migrations to use `get_or_create`
for the new model ensures that the external URL is unique, and we'll only be
copying over the first strain with that link, and updating all references to
that link to the same record.