Update Jenkinsfile to wait on ICE startup
The test suite has integration tests with ICE, and these depend on ICE being
up. This is normally not a problem, but with the change in storage backend,
tasks are running slowly sometimes, and it's possible for the EDD containers to
be up and ready to run tests, while ICE is still starting. This will cause all
of the test suite to error out (because of some session-scope fixtures).

Including a wait loop on the ICE container reporting healthy should prevent
issues with the tests all failing on first pass.