Fix ICE integration test
One of the ICE integration tests confirms that the "readonly" created user can
see strain entries. The integration tests create ten strains, but the user here
only can see the first five. The test is checking that the user can see ten or
more strains; this will fail on the first run, then succeed on every run
following. Fixing the test to check for five or more strains seen, so the
first-run test can succeed.