Remove numpy dependency
The only code in EDD using numpy was the interpolation function previously used
in SBML exports, which would infer a value for some measurement at a specific
timepoint in the SBML output, when EDD did not have any value for that
timepoint, but did have values before and after. This wasn't even really
desired by modelers, as it was doing a straight linear interpolation, and users
of the model had requested the ability to use a different interpolation method.
Regardless, the SBML export is removed, and the functionality for it likely
belongs outside of EDD itself.

As this was the only place that used numpy, it can be removed from the
dependencies list. This makes the overall image size about 64MB smaller.

Note, the libsbml dependency still exists and must be maintained until the
database migrations can be re-squashed and the import of libsbml in
`server/main/migrations/0001_edd_2_7.py` can be removed, as there should no
longer be a need to create the SBML models in a squashed migration. Doing
migration squashes requires some coordination of upgrades to EDD, and should
have additional documentation on specific commits to upgrade, preparing the
database, then continuing upgrade to the latest version.